Move fixtures to streaming_utils/conftest
