+++ source/streaming/tests/streaming_utils/input_source_readers/test_time_series_reader.py
@@ -1,0 +1,26 @@
+# Copyright 2020 Energinet DataHub A/S
+#
+# Licensed under the Apache License, Version 2.0 (the "License2");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import pytest
+from geh_stream.streaming_utils.input_source_readers.time_series_reader import parse_stream
+
+
+def test_parse_stream(event_hub_message_df):
+
+    parsed_time_series_point_stream = parse_stream(event_hub_message_df)
+
+    metering_point_id = "meteringpointid1"
+    first = parsed_time_series_point_stream.first()
+
+    assert first.series_meteringPointId == metering_point_id
